requirements.txt
Needed packages in python:

streamlit
pandas
numpy
torch
joblib
plotly
seaborn
matplotlib
python-dotenv
scikit-learn